+++ jenkins-plugin/src/test/java/org/jenkinsci/plugins/pipeline/maven/db/PipelineMavenPluginMySqlDaoIT.java
@@ -22,24 +22,37 @@
  * THE SOFTWARE.
  */
 
-package org.jenkinsci.plugins.pipeline.maven.dao;
-
-import static org.assertj.core.api.Assertions.assertThat;
+package org.jenkinsci.plugins.pipeline.maven.db;
 
 import javax.sql.DataSource;
 
-import org.h2.jdbcx.JdbcConnectionPool;
+import org.jenkinsci.plugins.pipeline.maven.db.AbstractPipelineMavenPluginDao;
+import org.jenkinsci.plugins.pipeline.maven.db.PipelineMavenPluginDaoAbstractTest;
+import org.jenkinsci.plugins.pipeline.maven.db.PipelineMavenPluginMySqlDao;
 import org.jenkinsci.plugins.pipeline.maven.db.migration.MigrationStep;
-import org.junit.jupiter.api.Test;
+import org.testcontainers.containers.MySQLContainer;
+import org.testcontainers.junit.jupiter.Container;
+import org.testcontainers.junit.jupiter.Testcontainers;
+
+import com.zaxxer.hikari.HikariConfig;
+import com.zaxxer.hikari.HikariDataSource;
 
 /**
  * @author <a href="mailto:cleclerc@cloudbees.com">Cyrille Le Clerc</a>
  */
-public class PipelineMavenPluginMySqlDaoTest extends PipelineMavenPluginDaoAbstractTest {
+@Testcontainers(disabledWithoutDocker = true)
+public class PipelineMavenPluginMySqlDaoIT extends PipelineMavenPluginDaoAbstractTest {
+
+    @Container
+    public static MySQLContainer<?> DB = new MySQLContainer<>(MySQLContainer.NAME);
 
     @Override
     public DataSource before_newDataSource() {
-        return JdbcConnectionPool.create("jdbc:h2:mem:;MODE=MYSQL", "sa", "");
+        HikariConfig config = new HikariConfig();
+        config.setJdbcUrl(DB.getJdbcUrl());
+        config.setUsername(DB.getUsername());
+        config.setPassword(DB.getPassword());
+        return new HikariDataSource(config);
     }
 
     @Override
@@ -60,22 +73,5 @@ public class PipelineMavenPluginMySqlDaoTest extends PipelineMavenPluginDaoAbstr
                 };
             }
         };
-    }
-
-    @Test
-    public void test_mariadb_version_parsing_JENKINS_55378() {
-        String actual = PipelineMavenPluginMySqlDao.extractMariaDbVersion("5.5.5-10.2.20-MariaDB");
-        assertThat(actual).isEqualTo("10.2.20");
-    }
-
-    /**
-     * docker run -e MYSQL_ROOT_PASSWORD=mypass -e MYSQL_DATABASE=jenkins -e
-     * MYSQL_USER=jenkins -e MYSQL_PASSWORD=jenkins -p 3307:3306 -d
-     * mariadb/server:latest
-     */
-    @Test
-    public void test_mariadb_version_parsing_mariadb_as_docker_container() {
-        String actual = PipelineMavenPluginMySqlDao.extractMariaDbVersion("5.5.5-10.3.11-MariaDB-1:10.3.11+maria~bionic");
-        assertThat(actual).isEqualTo("10.3.11");
     }
 }
