+++ pipeline-maven-api/src/main/resources/org/jenkinsci/plugins/pipeline/maven/dao/Messages.properties
@@ -1,0 +1,1 @@
+dao.null.description=Pipeline Maven Plugin no storage mode
